add encryption to passwords in the database
This is a rather crude way to do this - it's not really convenient for the user and the password will be stored in the shell history. However,
having looked at the alternatives, this is one of the better ones. The best option would of course be to have a daemon but that's a lot more work.