add ability to check master password integrity by comparing hash
This is of course superficial for security since the hash can be changed in the file but it still prevents accidental entries of the wrong password.
Without this, passwords in the database can be encrypted with the wrong password and the user will not know until they try to decrypt them.