add a config file to store password hashes
I realized that we almost certainly will have different hashes for different database paths so this makes
sure that works as expected. I also considered using a database for this but having a simple JSON file makes
it easier for the user to see and also allows me to add more actual configuration options in the future.